requirements.txt
fastapi
uvicorn
requests
python-dotenv
PyGithub
openai

